First working version with public key auth

+++ Sources/SpokeServer/Routes.swift
@@ -7,66 +7,112 @@ import JWTKit
 // SCRAM-SHA-256 doesn't require the password to leave the client's device
 // It's also more secure, since the server is being verified as well
 // TODO: Use $set queries to prevent simultanious updates to the same model, overwriting one another
-struct Credentials: Codable {
+struct Credentials: Content {
+    static let defaultContentType = HTTPMediaType.bson
+    
     let username: String
     let password: String
     let publicKeys: Document
 }
 
-struct ChangeCredentials: Codable {
-    let oldPassword: String
-    let newPassword: String
+struct FirstLoginRequest: Content {
+    static let defaultContentType = HTTPMediaType.bson
+    
+    let user: Reference<User>
+    let deviceId: ObjectId
+    let password: String
+}
+
+struct FirstLoginResponse: Content {
+    static let defaultContentType = HTTPMediaType.bson
+    
+    let token: String
+}
+
+struct SetupAccount: Content {
+    static let defaultContentType = HTTPMediaType.bson
+    
+    let identity: PublicSigningKey
 }
 
 struct AuthResponse: Content {
+    static let defaultContentType = HTTPMediaType.bson
+    
     let profile: UserProfile
     let token: String
 }
 
-func registerRoutes(to routes: RoutesBuilder) {
-    routes.post("auth", "register") { req -> EventLoopFuture<AuthResponse> in
-        let credentials = try req.content.decode(Credentials.self, using: BSONDecoder())
-        
-        let user = try User(username: credentials.username, password: credentials.password, publicKeys: credentials.publicKeys)
-        let token = try Application.signer.sign(Token(user: user, validFor: 24 * 3600))
-        
-        let authResponse = AuthResponse(profile: UserProfile(representing: user), token: token)
-        return user.save(in: req.meow).transform(to: authResponse)
-    }
+struct UpdateDeviceConfigRequest: Content {
+    static let defaultContentType = HTTPMediaType.bson
     
-    routes.post("auth", "login") { req -> EventLoopFuture<AuthResponse> in
-        let credentials = try req.content.decode(Credentials.self)
-        
-        return Reference<User>(unsafeTo: credentials.username).resolve(in: req.meow).flatMapThrowing { user in
-            try user.authenticate(credentials.password)
-            let token = try Application.signer.sign(Token(user: user, validFor: 24 * 3600))
+    let config: UserDeviceConfig
+}
+
+struct UserKeysResponse: Content {
+    static let defaultContentType = HTTPMediaType.bson
+    
+    let user: UserProfile
+    let devices: [UserDevice]
+}
+
+func registerRoutes(to routes: RoutesBuilder) {
+    #if Xcode
+    routes.get("reset") { req in
+        return app.meow.raw.drop().flatMapThrowing { () -> Response in 
+            for i in 0..<10 {
+                let user = try User(username: "test\(i)", password: "test\(i)", identity: nil)
+                _ = user.save(in: app.meow)
+            }
             
-            return AuthResponse(profile: UserProfile(representing: user), token: token)
+            return Response(status: .ok)
+        }
+    }
+    #endif
+    
+    routes.post("auth", "first-login") { req -> EventLoopFuture<FirstLoginResponse> in
+        let login = try req.content.decode(FirstLoginRequest.self, using: BSONDecoder())
+        
+        return login.user.resolve(in: req.meow).flatMap { user in
+            do {
+                try user.authenticate(login.password)
+                
+                let device = try UserDevice(_id: .init(UserDeviceId(user: login.user, device: login.deviceId)))
+                
+                let token = Token(
+                    device: device.$_id,
+                    exp: ExpirationClaim(value: .init(timeIntervalSinceNow: 3600))
+                )
+                
+                let signedToken = try Application.signer.sign(token)
+                
+                return device.save(in: req.meow).transform(to: FirstLoginResponse(token: signedToken))
+            } catch {
+                return req.eventLoop.makeFailedFuture(error)
+            }
         }
     }
     
-    let routes = routes.grouped(TokenAuthenticationMiddleware())
-    
     // TODO: Use $set instead of `save`
-    routes.post("auth", "change-password") { req -> EventLoopFuture<UserProfile> in
-        guard let currentUser = req.user else {
+    let protectedRoutes = routes.grouped(TokenAuthenticationMiddleware())
+    
+    protectedRoutes.post("auth", "setup") { req -> EventLoopFuture<UserProfile> in
+        guard let currentUser = req.userId else {
             throw Abort(.unauthorized)
         }
         
-        let changeCredentials = try req.content.decode(ChangeCredentials.self, using: BSONDecoder())
+        let setup = try req.content.decode(SetupAccount.self, using: BSONDecoder())
         
         return currentUser.resolve(in: req.meow).flatMapThrowing { user -> User in
             var user = user
-            try user.authenticate(changeCredentials.oldPassword)
-            try user.changePassword(to: changeCredentials.newPassword)
+            try user.changeIdentity(to: setup.identity)
             return user
         }.flatMap { user in
             user.save(in: req.meow).transform(to: UserProfile(representing: user))
         }
     }
     
-    routes.get("users", ":id") { req -> EventLoopFuture<UserProfile> in
-        guard let user = req.parameters.get("id", as: Reference<UserProfile>.self) else {
+    protectedRoutes.get("users", ":userId") { req -> EventLoopFuture<UserProfile> in
+        guard let user = req.parameters.get("userId", as: Reference<UserProfile>.self) else {
             throw Abort(.notFound)
         }
         
@@ -74,8 +120,8 @@ func registerRoutes(to routes: RoutesBuilder) {
     }
     
     // TODO: Use update op with $push
-    routes.post("users", ":id", "block") { req -> EventLoopFuture<UserProfile> in
-        guard let currentUser = req.user, let otherUser = req.parameters.get("id", as: Reference<User>.self) else {
+    protectedRoutes.post("users", ":userId", "block") { req -> EventLoopFuture<UserProfile> in
+        guard let currentUser = req.userId, let otherUser = req.parameters.get("userId", as: Reference<User>.self) else {
             throw Abort(.unauthorized)
         }
         
@@ -89,8 +135,8 @@ func registerRoutes(to routes: RoutesBuilder) {
     }
     
     // TODO: Use update op with $pull
-    routes.post("users", ":id", "unblock") { req -> EventLoopFuture<UserProfile> in
-        guard let currentUser = req.user, let otherUser = req.parameters.get("id", as: Reference<User>.self) else {
+    protectedRoutes.post("users", ":userId", "unblock") { req -> EventLoopFuture<UserProfile> in
+        guard let currentUser = req.userId, let otherUser = req.parameters.get("userId", as: Reference<User>.self) else {
             throw Abort(.unauthorized)
         }
         
@@ -103,23 +149,114 @@ func registerRoutes(to routes: RoutesBuilder) {
         }
     }
     
-    routes.post("users", ":id", "messages", "send") { req -> EventLoopFuture<ChatMessage> in
-        guard let currentUser = req.user, let otherUser = req.parameters.get("id", as: Reference<User>.self) else {
+    protectedRoutes.get("users", ":userId", "keys") { req -> EventLoopFuture<UserKeysResponse> in
+        guard let user = req.parameters.get("userId", as: Reference<UserProfile>.self) else {
+            throw Abort(.notFound)
+        }
+        
+        return user.resolve(in: req.meow).flatMap { user in
+            return req.meow(UserDevice.self).find(where: "_id.user" == user._id && "config" != nil)
+                .allResults(failable: true)
+                .map { devices in
+                    UserKeysResponse(user: user, devices: devices)
+                }
+        }
+    }
+    
+    protectedRoutes.post("current-user", "devices", ":deviceId") { req -> EventLoopFuture<UserDevice> in
+        guard
+            var currentUserDevice = req.device,
+            let userIdentity = req.user?.identity
+        else {
             throw Abort(.unauthorized)
         }
         
+        let update = try req.content.decode(UpdateDeviceConfigRequest.self, using: BSONDecoder())
+        currentUserDevice.config = update.config
+        
+        try update.config.publicKey.verifySignature(signedBy: userIdentity)
+        
+        return currentUserDevice.save(in: req.meow)
+            .transform(to: currentUserDevice)
+    }
+    
+    protectedRoutes.post("users", ":userId", "send-message") { req -> EventLoopFuture<ChatMessage> in
+        guard
+            let currentUserDevice = req.device,
+            let otherUser = req.parameters.get("userId", as: Reference<User>.self)
+        else {
+            throw Abort(.unauthorized)
+        }
+        
+        let messageId = req.headers["X-Message-Id"].first
+        // TODO: Push type
+        
         return otherUser.resolve(in: req.meow).flatMap { otherUser in
-            if otherUser.blockedUsers.contains(currentUser) {
+            if otherUser.blockedUsers.contains(currentUserDevice.$_id.user) {
                 return req.eventLoop.makeFailedFuture(Abort(.forbidden))
             }
             
             return req.body.collect()
                 .unwrap(or: Abort(.badRequest))
                 .flatMap { buffer in
+                    return req.meow(UserDevice.self)
+                        .find(where: "_id.user" == otherUser._id)
+                        .allResults(failable: true)
+                        .flatMap { devices in
+                            let devices = devices.map(\.$_id.device)
+                            let message = ChatMessage(
+                                _id: messageId,
+                                message: Document(buffer: buffer),
+                                from: currentUserDevice*,
+                                to: otherUser*,
+                                devices: Set(devices)
+                            )
+                            
+                            return message.save(in: req.meow).transform(to: message)
+                    }
+                }
+        }
+    }
+    
+    protectedRoutes.post("users", ":userId", "devices", ":deviceId", "send-message") { req -> EventLoopFuture<ChatMessage> in
+        guard
+            let currentUserDevice = req.device,
+            let otherUser = req.parameters.get("userId", as: Reference<User>.self),
+            let otherUserDeviceId = req.parameters.get("deviceId", as: ObjectId.self)
+        else {
+            throw Abort(.unauthorized)
+        }
+        
+        let messageId = req.headers["X-Message-Id"].first
+        // TODO: Push type
+        
+        let otherUserDevice = try Reference<UserDevice>(
+            unsafeToEncoded: UserDeviceId(
+                user: otherUser,
+                device: otherUserDeviceId
+            )
+        )
+        
+        return otherUser.resolve(in: req.meow).flatMap { otherUser in
+            if otherUser.blockedUsers.contains(currentUserDevice.$_id.user) {
+                return req.eventLoop.makeFailedFuture(Abort(.forbidden))
+            }
+            
+            return otherUserDevice.exists(in: req.meow).flatMapThrowing { exists in
+                if !exists {
+                    throw Abort(.notFound)
+                }
+            }
+        }.flatMap {
+            return req.body.collect()
+                .unwrap(or: Abort(.badRequest))
+                .flatMap { buffer in
                     let message = ChatMessage(
+                        _id: messageId,
                         message: Document(buffer: buffer),
-                        from: currentUser,
-                        to: otherUser*
+                        from: currentUserDevice*,
+                        to: otherUser,
+                        devices: [otherUserDeviceId]
                     )
                     
                     return message.save(in: req.meow).transform(to: message)
@@ -127,8 +264,8 @@ func registerRoutes(to routes: RoutesBuilder) {
         }
     }
     
-    routes.webSocket("messages", "watch") { req, websocket in
-        guard let user = req.user else {
+    protectedRoutes.webSocket("websocket") { req, websocket in
+        guard let user = req.userId, let device = req.device else {
             _ = websocket.close(code: .unexpectedServerError)
             return
         }
@@ -140,8 +277,23 @@ func registerRoutes(to routes: RoutesBuilder) {
             .sequentialForEach { message in
                 let bson = try BSONEncoder().encode(message)
                 let promise = req.eventLoop.makePromise(of: Void.self)
+                
+                guard message.devices.contains(where: { $0 == device.$_id.device }) else {
+                    // Skip this one, not intended for this device
+                    promise.succeed(())
+                    return promise.futureResult
+                }
+                
+                // Emit to all devices, this is just one
+                // TODO: $push
+                var message = message
+                message.devicesReceived.insert(device.$_id.device)
+                
                 websocket.send(raw: bson.makeData(), opcode: .binary, promise: promise)
-                return promise.futureResult.flatMap {
+                
+                return message.save(in: req.meow).flatMap { _ in
+                    promise.futureResult
+                }.flatMap {
                     // TODO: Multi-Device Support?
                     // TODO: What is a message never arrives at the client? Do we use acknowledgements (which become a receive-receipt)?
                     chatMessages.deleteOne(where: "_id" == message._id).transform(to: ())
@@ -150,33 +302,48 @@ func registerRoutes(to routes: RoutesBuilder) {
         
         emittingOldMessages.flatMap {
             chatMessages.buildChangeStream {
-                match("recipient" == user.reference)
+                match("fullDocument.recipient" == user.reference)
             }
-        }.whenComplete { result in
-            switch result {
-            case .success(let changeStream):
-                changeStream.forEach { notification in
-                    guard notification.operationType == .insert, let message = notification.fullDocument else {
-                        return true
+        }.map { changeStream in
+            changeStream.forEach { notification in
+                guard notification.operationType == .insert, let message = notification.fullDocument else {
+                    return true
+                }
+                
+                guard message.devices.contains(where: { $0 == device.$_id.device }) else {
+                    // Skip this one, not intended for this device
+                    return true
+                }
+                
+                do {
+                    var message = message
+                    message.devicesReceived.insert(device.$_id.device)
+                    let promise = req.eventLoop.makePromise(of: Void.self)
+                    let bson = try BSONEncoder().encode(message)
+                    
+                    websocket.send(raw: bson.makeData(), opcode: .binary, promise: promise)
+                    
+                    _ = message.save(in: req.meow).flatMap { _ in
+                        promise.futureResult
+                    }.flatMap {
+                        // TODO: Multi-Device Support?
+                        // TODO: What is a message never arrives at the client? Do we use acknowledgements (which become a receive-receipt)?
+                        chatMessages.deleteOne(where: "_id" == message._id).transform(to: ())
                     }
                     
-                    do {
-                        let bson = try BSONEncoder().encode(message)
-                        let promise = req.eventLoop.makePromise(of: Void.self)
-                        websocket.send(raw: bson.makeData(), opcode: .binary, promise: promise)
-                        _ = promise.futureResult.flatMap {
-                            chatMessages.deleteOne(where: "_id" == message._id).transform(to: ())
-                        }
-                        return true
-                    } catch {
-                        _ = websocket.close(code: .unexpectedServerError)
-                        return false
-                    }
+                    return true
+                } catch {
+                    _ = websocket.close(code: .unexpectedServerError)
+                    return false
                 }
-            case .failure:
-                _ = websocket.close(code: .unexpectedServerError)
             }
+        }.whenFailure { error in
+            _ = websocket.close(code: .unexpectedServerError)
         }
+        
+//        websocket.onBinary { websocket, buffer in
+//            let message = Document(buffer: buffer, isArray: false)
+//        }
     }
     
     // TODO: Create and manage group chat configs
